add async/await

+++ src/Blog.Model/Db/ArticleImage.cs
@@ -9,7 +9,7 @@ namespace Blog.Model.Db
         /// 对应文章ID
         /// </summary>
         [SugarColumn(ColumnName = "article_id")]
-        public int ArticleId { get; set; }
+        public long ArticleId { get; set; }
 
         /// <summary>
         /// 图片链接
